+++ CHANGELOG.md
@@ -11,6 +11,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 - Added another temporary variabel `tmpIndex` to store the current array index so that the `tmpInts` array can be used without limitations
 - Added support for arrays as function return type
 - Added support for `for each` loop 
+- Added native function `strSplit` #6
 
 ### Changed